refactor(rls): fold RLS into rbac.cpp; precompute BYPASSRLS on RoleClosure
rls.cpp/rls.h are gone. Both entry points were called only from
CollectAndEnforce, so after the move they are file-local and the header has no
reason to exist. One file now holds the access check and the plan rewrite it
performs.

BypassesRls no longer looks the acting Role up in the snapshot per governed
relation. BYPASSRLS is the start role's own attribute, exactly like
is_superuser, so RoleClosure precomputes it in ComputeRoleClosure -- from the
GetObject<Role> call that was already happening there for is_superuser.

  return closure.is_superuser || closure.bypasses_rls ||
         (!forced && closure.Owns(table));

Every term is load-bearing. Owns() already folds superuser in, but FORCE ROW
LEVEL SECURITY subjects the owner to policies while never subjecting a
superuser, so is_superuser has to be tested on its own. And Can() is not usable
here despite folding both: it is "owns OR was granted the privilege", so it
would exempt every role holding SELECT -- the population RLS exists to filter.

Verified the FORCE case directly: superuser-owner sees all rows under FORCE, as
does a BYPASSRLS role.
