+++ server/auth/role_closure.cpp
@@ -112,6 +112,7 @@ RoleClosure ComputeRoleClosure(const catalog::Snapshot& snapshot,
   // The superuser bit is the start role's own attribute -- never inherited.
   if (auto start = snapshot.GetObject<catalog::Role>(role)) {
     out.is_superuser = start->IsSuperuser();
+    out.bypasses_rls = start->Has(catalog::RoleOption::BypassRls);
   }
   return out;
 }
